% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdh.aurisano.check.plot.R
\name{bmdh.aurisano.check.plot}
\alias{bmdh.aurisano.check.plot}
\title{bmdh.aurisano.check.plot}
\usage{
bmdh.aurisano.check.plot(
  to.file = FALSE,
  toxval.db = "res_toxval_v95",
  sys.date = Sys.Date()
)
}
\arguments{
\item{to.file}{If TRUE, send the plot to a file}

\item{toxval.db}{Database version}

\item{sys.date}{The date of the database export}
}
\value{
None; writes plot to PDF
}
\description{
Plot the difference between the Aurisano and current BMDh values
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{labs}}, \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{ggtheme}}, \code{\link[ggplot2]{scale_continuous}}, \code{\link[ggplot2]{ggsave}}
 \code{\link[grDevices]{dev}}
}
