% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdh.cumdist.plot.R
\name{bmdh.cumdist.plot}
\alias{bmdh.cumdist.plot}
\title{Plot the cumulative distribution of number of studies}
\usage{
bmdh.cumdist.plot(
  to.file = F,
  toxval.db = "res_toxval_v95",
  sys.date = "2024-02-28"
)
}
\arguments{
\item{to.file}{If TRUE, send the plot to a file}

\item{toxval.db}{Database version}

\item{sys.date}{The date of the database export}
}
\value{
writes a plot to a file
}
\description{
`bmdh.cumdist.plot` plots the cumulative distribution of number of studies
}
