% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dcap_critical_effect_category_mappings.R
\name{check_dcap_toxicological_effect_category_mappings}
\alias{check_dcap_toxicological_effect_category_mappings}
\title{check_dcap_toxicological_effect_category_mappings}
\usage{
check_dcap_toxicological_effect_category_mappings(
  toxval.db,
  get_suggestions = TRUE,
  input_file = "",
  output_dir = ""
)
}
\arguments{
\item{toxval.db}{The version of ToxVal to use}

\item{get_suggestions}{Whether to provide mapping suggestions (Default: TRUE)}

\item{input_file}{The file to pull missing toxicological_effect_category from.}

\item{output_dir}{The folder used to write output to}
}
\value{
None; output is written to Excel files
}
\description{
Attempts to remap previously-mapped toxicological_effect_category values to DCAP entries
}
\details{
The output Excel files are as follows:
- dcap_mappings_identified.xlsx: All toxicological_effect_categories that could be confidently remapped
- dcap_mappings_still_missing.xlsx: Full data for entries missing categorizations
- dcap_missing_categorization.xlsx: Just toxicological_effect, study_type values missing categorizations
- dcap_mapping_suggestions.xlsx: Mapping suggestions based on close, but not exact, matches
}
