% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.fields.to.json.R
\name{convert.fields.to.json}
\alias{convert.fields.to.json}
\title{convert.fields.to.json}
\usage{
convert.fields.to.json(in_dat)
}
\arguments{
\item{in_dat}{Dataframe to use to generate the new "record" JSON field.}
}
\value{
Modified dataframe with new "record" field in JSON format.
}
\description{
Generate a new "record" field as a JSON key-value dictionary of row field values.
}
\seealso{
\code{\link[dplyr]{summarise}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{bind}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
}
