% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_get_conversion_factor.R
\name{convert_get_conversion_factor}
\alias{convert_get_conversion_factor}
\title{convert_get_conversion_factor}
\usage{
convert_get_conversion_factor(conv_factor = 1)
}
\arguments{
\item{conv_factor}{Conversion factor to use (such as Molecular weight, tissue Density, etc.)}
}
\value{
List of conversion factors
}
\description{
A helper function to convert input values to desired units.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
