% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_effect_source.R
\name{critical_effect_source}
\alias{critical_effect_source}
\title{Find the source for odfd critical effect chunks.}
\usage{
critical_effect_source(
  toxval.db = "res_toxval_v95",
  user = "_dataminer",
  password = "pass"
)
}
\arguments{
\item{toxval.db}{Database version}

\item{user}{The username for the MySQL database. The database instance is
hard-coded in the function setDBConn().}

\item{password}{The user's MySQL database password.}
}
\description{
Find the source for odfd critical effect chunks.
}
