% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcap.counts.R
\name{dcap.counts}
\alias{dcap.counts}
\title{dcap.counts}
\usage{
dcap.counts(toxval.db = "res_toxval_v95", sys.date = Sys.Date())
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the export}
}
\value{
Write a file with the filtered results:ToxValDB for BMDh filtered {toxval.db} {sys.date}.xlsx
}
\description{
Calculate some stats for DCAP
}
\details{
Gets some statistics for the DCAP project off of the current ToxValDB export
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}, \code{\link[openxlsx]{write.xlsx}}
}
