% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{driver}
\usage{
driver(toxval.db = "res_toxval_v95", sys.date = Sys.Date(), run.export = TRUE)
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the database export}

\item{run.export}{Whether to run the export.for.bmdh function (Default: TRUE)}
}
\value{
None
}
\description{
Run all of the calculations to go from database export to calculation of final BMDh values
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
