% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.for.bmdh.R
\name{export.for.bmdh}
\alias{export.for.bmdh}
\title{export.for.bmdh}
\usage{
export.for.bmdh(
  toxval.db = "res_toxval_v95",
  include.pesticides = FALSE,
  run_name = Sys.Date()
)
}
\arguments{
\item{toxval.db}{Database version}

\item{include.pesticides}{Flag to include pesticides in output or not}

\item{run_name}{The desired name for the output directory (Default: current date)}
}
\value{
Write a file with the results: ToxValDB for BMDh {toxval.db} {Sys.Date()}.xlsx
}
\description{
Export records required for calculating BMDh values.
}
\details{
Exports all of the data required for the BMDh calculations.
The main query may need to be modified to extract more columns if needed for
the final application. Certain sources have been excluded because they have a high
percentage of read-across values. Species are filtered to only include Human,
Dog, Mouse, Rat and Rabbit. If more species are to be included, then allometric
scaling factors for those need to added to the function bmd.per.study().
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{createStyle}}, \code{\link[openxlsx]{write.xlsx}}
}
