% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.for.critical_effect_mapping.R
\name{export.for.critical_effect_mapping}
\alias{export.for.critical_effect_mapping}
\title{Export records required for managing the critical effect categories for the BMD calculations.}
\usage{
export.for.critical_effect_mapping(
  toxval.db = "res_toxval_v95",
  user = "rjudson",
  password
)
}
\arguments{
\item{toxval.db}{Database version}

\item{user}{The username for the MySQL database. The database instance is
hard-coded in the function setDBConn().}

\item{password}{The user's MySQL database password.}
}
\value{
Write a file with the results: ToxValDB for BMDh {toxval.db} {Sys.Date()}.xlsx
}
\description{
`export.for.critical_effect_mapping` Exports all of the data required for
performing teh criticla effect mapping for the BMDh calculations
}
