% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.for.dcap.qc.R
\name{export.for.dcap.qc}
\alias{export.for.dcap.qc}
\title{Export records required for performing QC on the ToxValDB DCAP sources.}
\usage{
export.for.dcap.qc(
  toxval.db = "res_toxval_v95",
  user = "_dataminer",
  password = "pass"
)
}
\arguments{
\item{toxval.db}{Database version}

\item{user}{The username for the MySQL database. The database instance is
hard-coded in the function setDBConn().}

\item{password}{The user's MySQL database password.}
}
\value{
Write a file with the results: ToxValDB for BMDh {toxval.db} {Sys.Date()}.xlsx
}
\description{
`export.for.dcap.qc` Exports all of the data required for the doing DCAP QC calculations.
}
