% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.for.dcap.R
\name{export.for.dcap}
\alias{export.for.dcap}
\title{export.for.dcap}
\usage{
export.for.dcap(
  toxval.db,
  include.pesticides = FALSE,
  include.drugs = FALSE,
  include.epa_dws = TRUE,
  include.food_add = FALSE,
  run_name = Sys.Date(),
  db.type = db.type
)
}
\arguments{
\item{toxval.db}{Database version.}

\item{include.pesticides}{Flag to include pesticides in output or not, default FALSE.}

\item{include.drugs}{Flag to include drugs in output or not, default FALSE.}

\item{include.epa_dws}{Flag to include EPA DWS in output or not, default TRUE.}

\item{include.food_add}{Flag to include food additives in output or not, default FALSE.}

\item{run_name}{The desired name for the output directory, default current date.}

\item{db.type}{String of what kind of database connection to use, default "mysql. If "sqlite", workflow with use .Renv defined "sqlite_file" file path.}
}
\value{
Write a file with the results: ToxValDB for DCAP {toxval.db} {Sys.Date()}.xlsx
}
\description{
Export records required for DCAP.
}
\details{
None, exports all of the data required for the DCAP calculations as XLSX files.
}
\examples{
\dontrun{
if(interactive()){
 export.for.dcap(toxval.db = "res_toxval_v96_1")
 }
}
}
\seealso{
\code{\link[openxlsx]{createStyle}}, \code{\link[openxlsx]{write.xlsx}}
}
