% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.for.eco.qsar.R
\name{export.for.eco.qsar}
\alias{export.for.eco.qsar}
\title{Build a data frame of the data for the toxval manuscript}
\usage{
export.for.eco.qsar(toxval.db, source = NULL)
}
\arguments{
\item{toxval.db}{Database version}

\item{source}{The source to be updated}
}
\value{
Write a file with the results
}
\description{
Build a data frame of the data for the toxval manuscript
}
