% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.for.toxicological_effect_mapping.R
\name{export.for.toxicological_effect_mapping}
\alias{export.for.toxicological_effect_mapping}
\title{export.for.toxicological_effect_mapping}
\usage{
export.for.toxicological_effect_mapping(
  toxval.db = "res_toxval_v95",
  user = "rjudson",
  password
)
}
\arguments{
\item{toxval.db}{Database version}

\item{user}{The username for the MySQL database. The database instance is #' hard-coded in the function setDBConn().}

\item{password}{The user's MySQL database password.}
}
\value{
Write a file with the results: ToxValDB for BMDh {toxval.db} {Sys.Date()}.xlsx
}
\description{
Export records required for managing the critical effect categories for the BMD calculations.
}
\details{
Exports all of the data required for
performing the critical effect mapping for the BMDh calculations
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{createStyle}}, \code{\link[openxlsx]{write.xlsx}}
}
