% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.for.lel.R
\name{filter.for.lel}
\alias{filter.for.lel}
\title{Filter the exported records for redundancy}
\usage{
\method{filter}{for.lel}(toxval.db = "res_toxval_v95", sys.date = "2024-04-10")
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the export}
}
\value{
Write a file with the filtered results:ToxValDB for BMDh filtered {toxval.db} {sys.date}.xlsx
}
\description{
`filter.for.lel` Filters LEL, NEL values wihere a LOAEL/NOAEL value exists
}
