% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.pods.R
\name{filter.pods}
\alias{filter.pods}
\title{filter.pods}
\usage{
\method{filter}{pods}(toxval.db, run_name = Sys.Date())
}
\arguments{
\item{toxval.db}{Database name}

\item{run_name}{The desired name for the output directory (Default: current date)}
}
\value{
None; filtered results are recorded in Excel file
}
\description{
Filter values for BMDh according to specified POD rules
}
\details{
Filtering steps differ between authoritative and non-authoritative sources
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}, \code{\link[openxlsx]{createStyle}}, \code{\link[openxlsx]{write.xlsx}}
}
