% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pod.per.chemical.R
\name{filter.vals}
\alias{filter.vals}
\title{filter LOELs when NOELs are present for the same study
filter out redundant values for the same study group}
\usage{
\method{filter}{vals}(t2)
}
\description{
filter LOELs when NOELs are present for the same study
filter out redundant values for the same study group
}
