% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.study_group.R
\name{fix.study_group}
\alias{fix.study_group}
\title{Set the study_group field}
\usage{
fix.study_group(df)
}
\arguments{
\item{df}{Input dataframe to set study_group}
}
\value{
for each source writes an Excel file with the name
 ../export/export_by_source_{data}/toxval_all_{toxval.db}_{source}.xlsx
}
\description{
Set the study_group field
}
