% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chemical_index.R
\name{get_chemical_index}
\alias{get_chemical_index}
\title{get_chemical_index}
\usage{
get_chemical_index(input_toxval_file, input_qsar_file)
}
\arguments{
\item{input_toxval_file}{Filepath to input ToxVal file with DTXSID values.}

\item{input_qsar_file}{Filepath to input QSAR Ready Smiles file with DTXSID values.}
}
\value{
Combined chemical index dataframe. A CSV file is also written to the input subfolder.
}
\description{
Generate a chemical index based on ToxVal DTXSID list and QSAR Ready Smiles. The output file is reviewed and a final version is saved in the "data/input" folder.
}
\seealso{
\code{\link[readr]{read_delim}}, \code{\link[readr]{cols}}, \code{\link[readr]{write_delim}}
 \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{c("rowwise", "rowwise")}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{group_by}}
 \code{\link[purrr]{map}}
 \code{\link[digest]{digest}}
}
