% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_vars.R
\name{global_vars}
\alias{global_vars}
\title{global_vars}
\usage{
global_vars()
}
\value{
Named list of global variables.
}
\description{
Global variable list to use acros scripts.
}
\examples{
\dontrun{
if(interactive()){
 # Get vector of DCAP sources
 global_vars()$dcap_sources
 }
}
}
