% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxvaldbPodsEco.R
\name{pod.per.chemical.eco}
\alias{pod.per.chemical.eco}
\title{Calculate the eco PODs}
\usage{
pod.per.chemical.eco(toxval.db = "res_toxval_v95", sys.date = "2024-03-05")
}
\arguments{
\item{sys.date}{The date of the database export}

\item{toxvaldb}{The version fo the ToxValDB to use}
}
\description{
Calculate the eco PODs
}
