% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pod.rule.maker.R
\name{pod.rule.maker}
\alias{pod.rule.maker}
\title{pod.rule.maker}
\usage{
pod.rule.maker(
  toxval,
  hra.sources,
  rule.name = "Rule 3",
  rule.stype = "chronic",
  rule.ttype = "LO(A)EL, NO(A)EL, BMD",
  stype.list = c("chronic"),
  ttype.list = c(loel.types, noel.types, bmdl.types)
)
}
\arguments{
\item{toxval}{Input dataframe of ToxVal data.}

\item{hra.sources}{Human Risk Assessment source list}

\item{rule.name}{PARAM_DESCRIPTION, Default: 'Rule 3'}

\item{rule.stype}{PARAM_DESCRIPTION, Default: 'chronic'}

\item{rule.ttype}{PARAM_DESCRIPTION, Default: 'LO(A)EL, NO(A)EL, BMD'}

\item{stype.list}{PARAM_DESCRIPTION, Default: c("chronic")}

\item{ttype.list}{PARAM_DESCRIPTION, Default: c(loel.types, noel.types, bmdl.types)}
}
\value{
Modified version of input toxval dataframe.
}
\description{
Function to create/choose rule based on POD inputs
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{quantile}}, \code{\link[stats]{sd}}
}
