% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_final_output_file.R
\name{prep_final_output_file}
\alias{prep_final_output_file}
\title{prep_final_output_file}
\usage{
prep_final_output_file(toxval.db, run_name = Sys.Date())
}
\arguments{
\item{toxval.db}{Database version.}

\item{run_name}{The desired name for the output directory, default current date.}
}
\description{
Function to finalize prepped DCAP input file to remove intermediate fields and update field assignments.
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}
 \code{\link[writexl]{write_xlsx}}
}
