% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCurrentFunction.R
\name{printCurrentFunction}
\alias{printCurrentFunction}
\title{printCurrentFunction}
\usage{
printCurrentFunction(comment.string = NA)
}
\arguments{
\item{comment.string}{An optional string to be printed}
}
\value{
None. Console output generated.
}
\description{
Print the name of the current function to assist with logging and debugging.
}
\seealso{
\code{\link[utils]{flush.console}}
}
