% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pod.per.chemical.R
\name{rule.maker}
\alias{rule.maker}
\title{Make the rule}
\usage{
rule.maker(
  toxval,
  hra.sources,
  rule.name = "Rule 3",
  rule.stype = "chronic",
  rule.ttype = "LO(A)EL, NO(A)EL, BMD",
  stype.list = c("chronic"),
  ttype.list = c(loel.types, noel.types, bmdl.types)
)
}
\description{
Make the rule
}
