% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_toxvaldb_dcap_prep.R
\name{run_toxvaldb_dcap_prep}
\alias{run_toxvaldb_dcap_prep}
\title{run_toxvaldb_dcap_prep}
\usage{
run_toxvaldb_dcap_prep(
  toxval.db,
  run_name = Sys.Date(),
  run.export = TRUE,
  include.pesticides = FALSE,
  include.drugs = FALSE,
  include.epa_dws = TRUE,
  include.food_add = FALSE,
  db.type = "mysql"
)
}
\arguments{
\item{toxval.db}{Database name.}

\item{run_name}{The desired name for the output directory, default current date.}

\item{run.export}{Whether to run the export.for.dcap function, default TRUE.}

\item{include.pesticides}{Flag to include pesticides in output or not}

\item{include.drugs}{Flag to include drugs in output or not}

\item{include.epa_dws}{Flag to include EPA DWS in output or not}

\item{include.food_add}{Flag to include food additives in output or not}

\item{db.type}{String of what kind of database connection to use, default "mysql. If "sqlite", workflow with use .Renv defined "sqlite_file" file path.}
}
\value{
None, functions are run in sequence to generate output files.
}
\description{
Run all functions to pull records from ToxValDB and prep for DCAP analysis.
}
\examples{
\dontrun{
if(interactive()){
 run_toxvaldb_dcap_prep(toxval.db = "res_toxval_v96_1")
 }
}
}
