% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuery.R
\name{runQuery}
\alias{runQuery}
\title{runQuery}
\usage{
runQuery(query = NULL, db, do.halt = TRUE, verbose = FALSE, db.type = "mysql")
}
\arguments{
\item{query}{A properly formatted SQL query as a string.}

\item{db}{The name of the database.}

\item{do.halt}{If TRUE, halt on errors or warnings, default TRUE.}

\item{verbose}{If TRUE, print diagnostic information, default FALSE.}

\item{db.type}{String of what kind of database connection to use, default "mysql. If "sqlite", workflow with use .Renv defined "sqlite_file" file path.}
}
\value{
Dataframe of query results.
}
\description{
Runs a database query and returns a result set.
}
\examples{
\dontrun{
if(interactive()){
 runQuery(query = "SELECT * FROM toxval LIMIT 1", db = "res_toxval_v96_1")
 }
}
}
\seealso{
\code{\link[RMySQL]{character(0)}}, \code{\link[RMySQL]{MySQLDriver-class}}
 \code{\link[utils]{flush.console}}
}
