% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.species.list.R
\name{split.species.list}
\alias{split.species.list}
\title{Split species list}
\usage{
\method{split}{species.list}(df)
}
\arguments{
\item{df}{Input dataframe with species lists to split}
}
\value{
Modified input df with split species lists into multiple rows
}
\description{
Split species lists into multiple rows
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}
 \code{\link[tidyr]{separate_rows}}
 \code{\link[stringr]{str_trim}}, \code{\link[stringr]{case}}
}
