% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicological_effect_source.R
\name{toxicological_effect_source}
\alias{toxicological_effect_source}
\title{toxicological_effect_source}
\usage{
toxicological_effect_source(toxval.db = "res_toxval_v95", user, password)
}
\arguments{
\item{toxval.db}{Database version}

\item{user}{The username for the MySQL database. The database instance is #' hard-coded in the function setDBConn().}

\item{password}{The user's MySQL database password.}
}
\value{
Writes output to file
}
\description{
Find the source for odd critical effect chunks.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}, \code{\link[openxlsx]{write.xlsx}}
}
