% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.record.dedup.R
\name{toxval.record.dedup}
\alias{toxval.record.dedup}
\title{toxval.record.dedup}
\usage{
toxval.record.dedup(
  res,
  dedup_fields = NULL,
  hashing_cols = NULL,
  delim = " |::| "
)
}
\arguments{
\item{res}{Dataframe to process and collapse duplicate record fields.}

\item{dedup_fields}{vector containing field names to deduplicate, default NULL (all fields but hashing cols).}

\item{hashing_cols}{vector containing field names of hashing columns, default NULL.}

\item{delim}{String used to separate collapsed values, default ' |::| '.}
}
\value{
Dataframe containing deduplicated data, with duplicate records collapsed by delimiter in non-identifier fields.
}
\description{
Perform deduplication on data based on select identifier fields. Non-identifier fields will be collapsed based on the input delimiter.
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{na_if}}, \code{\link[dplyr]{distinct}}
}
