% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_hash_vectorized.R
\name{source_hash_vectorized}
\alias{source_hash_vectorized}
\title{source_hash_vectorized}
\usage{
source_hash_vectorized(res, hashing_cols)
}
\arguments{
\item{res}{The data frame to be processed}

\item{hashing_cols}{Optional list of columns to use for generating source_hash}
}
\value{
Input dataframe with new source_hash field
}
\description{
Generate the hash key for a source table based on hashing columns
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[digest]{digest}}
 \code{\link[dplyr]{distinct}}
}
