% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxvaldb.statplots.R
\name{toxvaldb.statplots}
\alias{toxvaldb.statplots}
\title{Plot statistics for ToxValDB for DCAP sources}
\usage{
toxvaldb.statplots(
  to.file = F,
  toxval.db = "res_toxval_v95",
  sys.date = "2024-04-16"
)
}
\arguments{
\item{dir}{The directory where the lists are stored}
}
\description{
Plot statistics for ToxValDB for DCAP sources
}
