% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pod.per.chemical.eco.R
\name{pod.per.chemical.eco}
\alias{pod.per.chemical.eco}
\title{pod.per.chemical.eco}
\usage{
pod.per.chemical.eco(toxval.db = "res_toxval_v95", sys.date = Sys.Date())
}
\arguments{
\item{toxval.db}{The version fo the ToxValDB to use}

\item{sys.date}{The date of the database export}
}
\value{
Writes output to file
}
\description{
Calculate the eco PODs
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}, \code{\link[openxlsx]{write.xlsx}}
 \code{\link[stats]{quantile}}, \code{\link[stats]{sd}}
}
