% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dcap_summary.R
\name{write_dcap_summary}
\alias{write_dcap_summary}
\title{write_dcap_summary}
\usage{
write_dcap_summary(toxval.db, sys.date = Sys.Date())
}
\arguments{
\item{toxval.db}{The version of ToxVal to use}

\item{sys.date}{The date of the export to use, Default: Sys.Date()}
}
\value{
None; writes summary reports to file
}
\description{
Output summary reports containing chemical and key field information
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[read_xlsx][readxl::read_xlsx]
 [mutate][dplyr::mutate], [filter][dplyr::filter], [select][dplyr::select], [distinct][dplyr::distinct], [left_join][dplyr::left_join], [pull][dplyr::pull], [ensym][dplyr::ensym]
 [replace_na][tidyr::replace_na]
 [tibble][tibble::tibble]
 [write.xlsx][openxlsx::write.xlsx]
}
