% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdh.cumdist.plot.R
\name{bmdh.cumdist.plot}
\alias{bmdh.cumdist.plot}
\title{bmdh.cumdist.plot}
\usage{
bmdh.cumdist.plot(
  to.file = F,
  toxval.db = "res_toxval_v95",
  sys.date = Sys.Date()
)
}
\arguments{
\item{to.file}{If TRUE, send the plot to a file}

\item{toxval.db}{Database version}

\item{sys.date}{The date of the database export}
}
\value{
writes a plot to a file
}
\description{
Plot the cumulative distribution of number of studies
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{labs}}, \code{\link[ggplot2]{stat_ecdf}}, \code{\link[ggplot2]{ggtheme}}, \code{\link[ggplot2]{lims}}, \code{\link[ggplot2]{ggsave}}
 \code{\link[ggpubr]{ggarrange}}
 \code{\link[grDevices]{dev}}
}
