% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcap.counts.R
\name{dcap.counts}
\alias{dcap.counts}
\title{Calculate some stats for DCAP}
\usage{
dcap.counts(toxval.db = "res_toxval_v95", sys.date = "2024-04-09")
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the export}
}
\value{
Write a file with the filtered results:ToxValDB for BMDh filtered {toxval.db} {sys.date}.xlsx
}
\description{
`dcap.counts` Gets some statistics for the DCAP project off of the current
ToxValDB export
}
