% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{driver}
\usage{
driver(
  toxval.db = "res_toxval_v95",
  run_name = Sys.Date(),
  run.export = TRUE,
  include.pesticides = FALSE
)
}
\arguments{
\item{toxval.db}{Database version}

\item{run_name}{The desired name for the output directory (Default: current date)}

\item{run.export}{Whether to run the export.for.bmdh function (Default: TRUE)}

\item{include.pesticides}{Flag to include pesticides in output or not}
}
\value{
None
}
\description{
Run all of the calculations to go from database export to calculation of final BMDh values
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
