% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{driver}
\usage{
driver(
  toxval.db = "res_toxval_v95",
  sys.date = Sys.Date(),
  user = "user",
  password = "password"
)
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the database export}

\item{user}{The username for the MySQL database. The database instance is #' hard-coded in the function setDBConn().}

\item{password}{The user's MySQL database password.}
}
\value{
None
}
\description{
Run all of the calculations to go from database export to calculation of final BMDh values
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
