% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{driver}
\usage{
driver(
  toxval.db,
  run_name = Sys.Date(),
  run.export = TRUE,
  include.pesticides = FALSE,
  include.drugs = FALSE,
  include.epa_dws = TRUE,
  include.food_add = FALSE,
  reset.study_group = FALSE
)
}
\arguments{
\item{toxval.db}{Database name}

\item{run_name}{The desired name for the output directory (Default: current date)}

\item{run.export}{Whether to run the export.for.bmdh function (Default: TRUE)}

\item{include.pesticides}{Flag to include pesticides in output or not}

\item{include.drugs}{Flag to include drugs in output or not}

\item{include.epa_dws}{Flag to include EPA DWS in output or not}

\item{include.food_add}{Flag to include food additives in output or not}

\item{reset.study_group}{Flag to reset study_group}
}
\value{
None
}
\description{
Run all of the calculations to go from database export to calculation of final BMDh values
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
