% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.for.eco.qsar.R
\name{export.for.eco.qsar}
\alias{export.for.eco.qsar}
\title{export.for.eco.qsar}
\usage{
export.for.eco.qsar(toxval.db, source = NULL)
}
\arguments{
\item{toxval.db}{Database version}

\item{source}{The source to be updated}
}
\value{
Write a file with the results
}
\description{
Build a data frame of the data for the toxval manuscript
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[utils]{write.table}}
 \code{\link[openxlsx]{createStyle}}, \code{\link[openxlsx]{write.xlsx}}
}
