% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.for.ecopods.R
\name{export.for.ecopods}
\alias{export.for.ecopods}
\title{Export records required for calculating POD values for ecotox QSAR models.
The data is exported as a xlsx and csv file because some records may cause the Excel format
to break. You can then open the csv (always works) and save as Excel}
\usage{
export.for.ecopods(
  toxval.db = "res_toxval_v95",
  do.load = T,
  user = "rjudson",
  password
)
}
\arguments{
\item{toxval.db}{Database version}

\item{do.load}{If TRUE, read from the database into a global}
}
\value{
Write a file with the results
}
\description{
Export records required for calculating POD values for ecotox QSAR models.
The data is exported as a xlsx and csv file because some records may cause the Excel format
to break. You can then open the csv (always works) and save as Excel
}
