% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.for.lel.R
\name{filter.for.lel}
\alias{filter.for.lel}
\title{filter.for.lel}
\usage{
\method{filter}{for.lel}(toxval.db = "res_toxval_v95", sys.date = Sys.Date())
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the export}
}
\value{
Write a file with the filtered results:ToxValDB for BMDh filtered {toxval.db} {sys.date}.xlsx
}
\description{
Filter the exported records for redundancy
}
\details{
Filters LEL, NEL values where a LOAEL/NOAEL value exists
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}, \code{\link[openxlsx]{createStyle}}, \code{\link[openxlsx]{write.xlsx}}
}
