% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.pods.R
\name{filter.pods}
\alias{filter.pods}
\title{filter.pods}
\usage{
\method{filter}{pods}(toxval.db, run_name = Sys.Date(), db.type)
}
\arguments{
\item{toxval.db}{Database name.}

\item{run_name}{The desired name for the output directory, default current date.}

\item{db.type}{String of what kind of database connection to use, default "mysql. If "sqlite", workflow with use .Renv defined "sqlite_file" file path.}
}
\value{
None, filtered results are recorded in Excel file.
}
\description{
Filter values for DCAP according to specified POD rules.
}
\examples{
\dontrun{
if(interactive()){
 filter.pods(toxval.db = "res_toxval_v96_1")
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}, \code{\link[openxlsx]{createStyle}}, \code{\link[openxlsx]{write.xlsx}}
}
