% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.summary.R
\name{filter.summary}
\alias{filter.summary}
\title{Summarize the filtering steps}
\usage{
\method{filter}{summary}(toxval.db = "res_toxval_v95", sys.date = "2024-04-10", do.load = T)
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the export}
}
\value{
Write a file with the filtered results:ToxValDB for BMDh LEL NEL multiNOEL filtered {toxval.db} {sys.date}.xlsx
}
\description{
`filter.for.multi.noel` Filters where multiple NOEL/NOEL etc. exist. For each study_group
this will select the highest NO(A)EL below the lowest LO(A)EL and the lowest LO(A)EL.
In all cases, all BMDx values are included
}
