% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/set.conceptual_model.by.critical_effect_category.R
\name{get.conceptual_model.by.critical_effect_category}
\alias{get.conceptual_model.by.critical_effect_category}
\title{get.conceptual_model.by.critical_effect_category}
\usage{
get.conceptual_model.by.critical_effect_category(df)
}
\arguments{
\item{df}{Input dataframe of study_type and critical_effect data.}
}
\value{
DataFrame map of models by critical_effect and study_type
}
\description{
Get the conceptual model based on critical_effect_category
}
