% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.current.run.directory.R
\name{init.current.run.directory}
\alias{init.current.run.directory}
\title{init.current.run.directory}
\usage{
init.current.run.directory(run_name = Sys.Date())
}
\arguments{
\item{run_name}{The desired name for the output directory, default current date.}
}
\value{
None. new folder directory is created.
}
\description{
Initialize output directory for current run.
}
