% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pod.per.chemical.R
\name{pod.per.chemical}
\alias{pod.per.chemical}
\title{Explore different methods for calculating PODs
This uses the same input as the function bmdh.perstudy / bmdh.per.chemical}
\usage{
pod.per.chemical(toxval.db = "res_toxval_v95", sys.date = "2024-03-05")
}
\arguments{
\item{toxval.db}{The version of ToxValDB to use}

\item{sys.date}{Date of the most recent data export}
}
\description{
Explore different methods for calculating PODs
This uses the same input as the function bmdh.perstudy / bmdh.per.chemical
}
