% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.counter.R
\name{property.counter}
\alias{property.counter}
\title{Count proerties of records}
\usage{
property.counter(toxval.db = "res_toxval_v95", sys.date = "2024-03-26")
}
\arguments{
\item{toxval.db}{Database version}

\item{user}{The username for the MySQL database. The database instance is
hard-coded in the function setDBConn().}

\item{password}{The user's MySQL database password.}
}
\value{
Write a file with the results: ToxValDB for BMDh {toxval.db} {Sys.Date()}.xlsx
}
\description{
`export.for.bmdh` Exports all of the data required for the BMDh calculations.
The main query may need to be modified to extract more columns if needed for
the final application. Certain sources have been excluded because they have a high
percentage of read-across values. Species are filtered to only include Human,
Dog, Mouse, Rat and Rabbit. If more species are to be included, then allometric
scaling factors for those need to added to the function bmd.per.study().
}
