% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.per.chemical.R
\name{studies.per.chemical}
\alias{studies.per.chemical}
\title{Calculate some stats for DCAP}
\usage{
studies.per.chemical(toxval.db = "res_toxval_v95", sys.date = "2024-04-03")
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the export}
}
\value{
Write a file with the filtered results:ToxValDB for BMDh filtered {toxval.db} {sys.date}.xlsx
}
\description{
`studies.per.chemical` Gets teh number of studies per chemical from the current
ToxValDB export
}
