% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_group.multichem.R
\name{study_group.multichem}
\alias{study_group.multichem}
\title{study_group.multichem}
\usage{
study_group.multichem(toxval.db, run_name = Sys.Date())
}
\arguments{
\item{toxval.db}{Database version}

\item{run_name}{The desired name for the output directory (Default: current date)}
}
\value{
Write a file with the filtered results:ToxValDB for BMDh LEL NEL multiNOEL filtered {toxval.db} {sys.date}.xlsx
}
\description{
Find study groups that span multiple chemicals
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}, \code{\link[openxlsx]{createStyle}}, \code{\link[openxlsx]{write.xlsx}}
}
