% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_hash_vectorized.R
\name{source_hash_vectorized}
\alias{source_hash_vectorized}
\title{source_hash_vectorized}
\usage{
source_hash_vectorized(res, hashing_cols)
}
\arguments{
\item{res}{The data frame to receive a hash key.}

\item{hashing_cols}{Optional list of columns to use for generating hash key field.}
}
\value{
Modified input dataframe with new source_hash field.
}
\description{
Generate the hash key for an input table based on hashing columns. The hash key helps identify duplicate records in a dataframe.
}
\seealso{
\code{\link[digest]{digest}}
 \code{\link[dplyr]{distinct}}
}
